implementazione add_file di UserFileManager
