+++ user_info.py
@@ -1,5 +1,4 @@
 import json, os
-from secret_stuff import private_folder
 from user_file_manager import UserFileManager
 
 # Racchiude tutte le informazioni di un utente
@@ -9,7 +8,6 @@ class UserInfo:
         self.id = id
         self.name = name
         self.state = state
-        self.files = files
 
         self.file_manager = UserFileManager(id, files)
 
@@ -21,46 +19,14 @@ class UserInfo:
             return self.id == other.id
 
     def add_file(self, user_file_name:str):
-        # Check if the private user directory exists
-        if not os.path.exists(self.get_user_folder()):
-            os.mkdir(self.get_user_folder())
-        
-        # Check if file already exists
-        if self.is_file(user_file_name):
-            # Cannot create file
-            return False
-        
-        file_opened = open(self.get_complete_file_name(user_file_name), "w")
-        file_opened.write(user_file_name)
-        file_opened.close()
-
-        self.files.append(user_file_name)
-
-        return True
-
-    def is_file(self, file_name:str) -> bool:
-        file_name = file_name.replace(' ', '_').lower()
-
-        for saved_file in self.files:
-            if (saved_file.replace(' ', '_').lower()) == file_name:
-                return True
-            
-        return False
-
-    def get_user_folder(self) -> str:
-        return f'{private_folder}{self.id}'
-    
-    def get_complete_file_name(self, user_file_name) -> str:
-        correct_file_name = user_file_name.replace(' ', '_').lower()
-
-        return f'{self.get_user_folder()}\\{correct_file_name}.json'
+        return self.file_manager.add_file(user_file_name)
 
     # Create a dictionary
     def user_to_table(self) -> dict:
         return { 
             'id': self.id, 
             'name': self.name,
-            'files' : self.files
+            'files' : self.file_manager.get_files()
         }
 
     # Json string
